Squashed commit of the following:
commit 517ffd0c0a61e468f2548fd464539cf8449a73dd
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sun Feb 1 17:33:36 2026 +0100

    Fix bytesPerRow calculation in writeTexture function

commit 1b6044c918a79f84f1ea0ffa1ac8ffe30e772cc6
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sat Jan 31 17:46:45 2026 +0100

    Fix alpha layer when gridRatio is not 1

commit 6a851cd14718a517db975930f07ef04496a4a68f
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sat Jan 31 17:29:00 2026 +0100

    Fix layer serialization to handle Uint8Array

commit 1c1618538cb1886e3b53fc98e1df5941c0e4441f
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sat Jan 31 17:15:52 2026 +0100

    Fix pan and zoom calculations for non-square grids

commit 7a8ebbe928789014868278981076226d70ba1714
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sat Jan 31 15:37:45 2026 +0100

    Fix zoom and pan

commit 3fb94c1878a3141cc3cf9c96bf38ab5c2111f0ce
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Fri Jan 30 18:03:41 2026 +0100

    Remove grid shader references from shader module and types

commit 48414c8e680ea4f0ff2a62209990b2e94f25e3ce
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Thu Jan 29 21:31:10 2026 +0100

    Add @webgpu/types to tsconfig

commit cc0c9f4adaeb17c2a4a751fd3cd9ce2dc5084e85
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Thu Jan 29 21:10:37 2026 +0100

    Fix zoom + pan

commit 0fd588b79ce0c342485d2bf64a1084190540f4aa
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Thu Jan 29 08:46:17 2026 +0100

    Prevent thickness from influencing selectionTool

commit fbf31687e6e16cba45ea644b98d8c676cb225db2
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Wed Jan 28 19:18:26 2026 +0100

    Add selectionToolActivated for ui buffer

commit 6efb418255c60d9ceb46e146454d857f304ee795
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Wed Jan 28 18:45:50 2026 +0100

    fix erase

commit 474657264a81e7fbde9055d5beda2a142e5bd529
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Wed Jan 28 18:29:39 2026 +0100

    add middleware handler

commit 00534debc09fdf7c3f285977610c6905a329c5b1
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Wed Jan 28 18:16:10 2026 +0100

    Remove unnecessary params from renderHandler

commit c99bf74dca3773c69ca57365caff0e2a1bb3c02e
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Wed Jan 28 18:13:31 2026 +0100

    Fix rect tool

commit 08320bf91b041dba74167c7311fd8b249ca1c27a
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Tue Jan 27 21:43:47 2026 +0100

    Fix erase

commit 35d9f02ca8314d84dc6fd86559f6374e6bca3568
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Tue Jan 27 21:38:12 2026 +0100

    Fix ui and paint position

commit 15fe279d6dbef3b0c2db41a9393b2a97ff5ee313
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Tue Jan 27 09:42:02 2026 +0100

    User uniform buffers from handler

commit 459d5ad75c15bf0d85e65aa2fe45a153d6fa3fd6
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Mon Jan 26 22:19:37 2026 +0100

    add uniformBuffersHandler

commit e7f94453cfa6a5f5e1abaa4ff205f3c53140b0b6
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Mon Jan 26 09:19:47 2026 +0100

    use uniformData for  ui shader

commit e731aa0632ed44100af8c10c87fc514ce6bcde11
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Mon Jan 26 08:55:48 2026 +0100

    remove old gridValues

commit 0e0c72e61540cae518b6462f9a5441d120b08a3c
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Mon Jan 26 08:50:37 2026 +0100

    add alpha layer to preview

commit 730e33220e527234cd9c27e403b77c61aee2fb7f
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sun Jan 25 18:19:30 2026 +0100

    Add alpha layer shaders

commit 12aaea5d0cd76aecd6aa006525a3c35967e4b919
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sat Jan 24 23:48:02 2026 +0100

    Fix preview to show layer texture shader

commit 6c591237fb61378118a8aa52b6b1ae2d61e3e731
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sat Jan 24 21:28:34 2026 +0100

    Update brush ui position and size to ajust to uint8

commit c5eb4403c62eaf1aa95888f8fb90ee0e13e7ecf0
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sat Jan 24 21:15:31 2026 +0100

    Use Uint8Array

commit e499c049a979ea23202526b585221d27ee7db720
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sat Jan 24 19:27:57 2026 +0100

    convert 32 to 8

commit 31a183039cdcae7d0e949dbddbc071ec2b882678
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sat Jan 24 14:56:55 2026 +0100

    convert to RRGGBBAA

commit 93a3bd528c21949b6499e8b2a3ac064df8d0b4e2
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Sat Jan 24 11:16:32 2026 +0100

    Add layertexture

commit ad6348936063f2c570b3372fe6b1c9ca3c49ad75
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Thu Jan 22 22:37:15 2026 +0100

    Use texture from buffer

commit 853ff65a6917f555404ddbeadb4834ea92cd93ba
Author: Igor Pieruccini <igorpieruccini@gmail.com>
Date:   Wed Jan 21 23:05:23 2026 +0100

    Improve render performance

+++ src/serialization/project.ts
@@ -10,7 +10,7 @@ export type SerializedProject = {
 };
 
 export const getSerializedBuffers = (
-  buffers: Map<string, Uint32Array<ArrayBuffer>>,
+  buffers: Map<string, Uint8Array<ArrayBuffer>>,
   gridSize: Vec2,
 ): Record<string, number[]> => {
   let serialized: Record<string, number[]> = {};
@@ -27,7 +27,7 @@ export const serialize = (
   name: string,
   gridSize: Vec2,
   layers: Layers,
-  buffers: Map<string, Uint32Array<ArrayBuffer>>,
+  buffers: Map<string, Uint8Array<ArrayBuffer>>,
 ): SerializedProject => {
   const serializedBuffers = getSerializedBuffers(buffers, gridSize);
