+++ src/storageDB.ts
@@ -25,27 +25,27 @@ export const localDataBase = async (name: string) => {
   const db = await openDB(name);
 
   async function save(
-    buffer: Uint32Array<ArrayBuffer>,
+    buffer: Uint8Array<ArrayBuffer>,
     layerId: string,
   ): Promise<void> {
     return new Promise((resolve, reject) => {
       const tx = db.transaction(`layers`, "readwrite");
       const store = tx.objectStore(`layers`);
-      const bufferCopy = new Uint32Array(buffer).slice();
+      const bufferCopy = new Uint8Array(buffer).slice();
       store.put(bufferCopy, layerId);
       tx.oncomplete = () => resolve();
       tx.onerror = (e) => reject((e.target as IDBRequest).error);
     });
   }
 
-  async function load(layerId: string): Promise<Uint32Array<ArrayBuffer>> {
+  async function load(layerId: string): Promise<Uint8Array<ArrayBuffer>> {
     return new Promise((resolve, reject) => {
       const tx = db.transaction(`layers`, "readonly");
       const store = tx.objectStore(`layers`);
       const request = store.get(layerId);
 
       request.onsuccess = (e) =>
-        resolve((e.target as IDBRequest).result as Uint32Array<ArrayBuffer>);
+        resolve((e.target as IDBRequest).result as Uint8Array<ArrayBuffer>);
       request.onerror = (e) => reject((e.target as IDBRequest).error);
     });
   }
